海南广告更换成popupwindow

+++ app/src/main/java/com/huicheng/hotel/android/ui/custom/CustomCirclePieChart.java
@@ -173,10 +173,10 @@ public class CustomCirclePieChart extends View {
             upDis = 10;
             downDis = 10;
         } else {
-            upDis = -linePaint.measureText(context.getResources().getString(R.string.RMB) + String.valueOf(cost)) - 10;
+            upDis = -linePaint.measureText(context.getResources().getString(R.string.RMB) + String.valueOf((int)cost)) - 10;
             downDis = -linePaint.measureText(strMethod) - 10;
         }
-        canvas.drawText(context.getResources().getString(R.string.RMB) + String.valueOf(cost), end2X + endX + upDis, endY - 15, linePaint);
+        canvas.drawText(context.getResources().getString(R.string.RMB) + String.valueOf((int)cost), end2X + endX + upDis, endY - 15, linePaint);
         canvas.drawText(strMethod, end2X + endX + downDis, endY + 30, linePaint);
     }
 }
